feat: refactor OOps struct to be truly build-in
