+++ H/entry.h
@@ -69,6 +69,7 @@
 #include "pitch.h"
 #include "opcode.h"
 #include "array_ops.h"
+#include "struct_ops.h"
 
 #define S(x)    sizeof(x)
