+++ OOps/array_ops.c
@@ -727,84 +727,6 @@ int32_t tabarithset2(CSOUND *csound, TABARITH2 *p)
   return OK;
 }
 
-
-int32_t array_set_struct(CSOUND* csound, ARRAY_SET *p)
-{
-  ARRAYDAT* dat = p->arrayDat;
-  if (UNLIKELY(dat == NULL || dat->data == NULL)) {
-    return csound->PerfError(csound, &(p->h), Str("array_set_struct: NULL array"));
-  }
-  int32_t indefArgCount = p->INOCOUNT - 2; // value + at least one index
-  if (UNLIKELY(indefArgCount <= 0))
-    return csound->PerfError(csound, &(p->h), Str("array_set_struct: no indexes"));
-  if (UNLIKELY(dat->dimensions > 0 && indefArgCount != dat->dimensions)) {
-    // Allow flat arrays with no metadata
-    if (!(dat && dat->dimensions == 0)) {
-      return csound->PerfError(csound, &(p->h), Str("array_set_struct: dimension mismatch"));
-    }
-  }
-  // Compute flat index
-  int32_t index = 0;
-  for (int32_t i = 0; i < indefArgCount; i++) {
-    int32_t end = (int32_t)(*p->indexes[i]);
-    if (dat->dimensions > 0 && dat->sizes != NULL) {
-      if (UNLIKELY(end < 0 || end >= dat->sizes[i])) {
-        return csound->PerfError(csound, &(p->h), Str("array_set_struct: index %d out of range"), i+1);
-      }
-      index = (index * dat->sizes[i]) + end;
-    } else {
-      index = (i == 0) ? end : (index * 0 + end);
-    }
-  }
-  // Address element in bytes (safe for struct payloads)
-  char* base = (char*)dat->data;
-  char* dstPtr = base + (index * (int32_t)dat->arrayMemberSize);
-
-  // Ensure destination struct is initialized (members allocated)
-  if (dat->arrayType && dat->arrayType->userDefinedType) {
-    CS_STRUCT_VAR* dst = (CS_STRUCT_VAR*)(void*)dstPtr;
-    if (dst && dst->members == NULL) {
-      CS_VARIABLE* var = dat->arrayType->createVariable(csound, (void*)dat->arrayType, p->h.insdshead);
-      if (var && var->initializeVariableMemory) {
-        var->initializeVariableMemory(csound, var, (MYFLT*)dst);
-      }
-    }
-  }
-
-  // Copy value into destination using type-aware copier
-  // For struct arrays, we need to copy member by member since the array elements
-  // are raw struct data, not CS_STRUCT_VAR structures with members pointers
-  if (dat->arrayType && dat->arrayType->userDefinedType) {
-    CS_STRUCT_VAR* srcVar = (CS_STRUCT_VAR*)p->value;
-    CS_STRUCT_VAR* dstVar = (CS_STRUCT_VAR*)dstPtr;
-
-    // Ensure both are initialized
-    if (srcVar && srcVar->members && dstVar && dstVar->members && dat->arrayType->members) {
-      // Get member count from the type definition, not the variable
-      int32_t memberCount = cs_cons_length(dat->arrayType->members);
-      for (int32_t i = 0; i < memberCount; i++) {
-        CS_VAR_MEM* srcMember = srcVar->members[i];
-        CS_VAR_MEM* dstMember = dstVar->members[i];
-        if (srcMember && dstMember && srcMember->varType) {
-          // Copy the member value
-          if (srcMember->varType->copyValue) {
-            srcMember->varType->copyValue(csound, srcMember->varType,
-                                         &dstMember->value, &srcMember->value, p->h.insdshead);
-          } else {
-            dstMember->value = srcMember->value;
-          }
-        }
-      }
-    }
-  } else if (dat->arrayType && dat->arrayType->copyValue) {
-    dat->arrayType->copyValue(csound, dat->arrayType, (void*)dstPtr, p->value, p->h.insdshead);
-  } else {
-    // Fallback should not happen for struct, but keep parity with array_set
-    if (LIKELY(dstPtr != NULL && p->value != NULL)) *((MYFLT*)dstPtr) = *((MYFLT*)p->value);
-  }
-  return OK;
-}
-
 int32_t tabadd(CSOUND *csound, TABARITH *p)
 {
   ARRAYDAT *ans = p->ans;
