+++ Top/csmodule.c
@@ -1210,7 +1210,6 @@ extern int32_t scugens_localops_init(CSOUND *, void *);
 extern int32_t emugens_localops_init(CSOUND *, void *);
 extern int32_t control_localops_init(CSOUND *, void *);
 extern int32_t urandom_localops_init(CSOUND *, void *);
-extern int32_t structops_localops_init(CSOUND *, void *);
 
 #ifdef HAVE_SOCKETS
 extern int32_t socksend_localops_init(CSOUND *, void *);
@@ -1314,7 +1313,7 @@ CS_NOINLINE int32_t csoundInitStaticModules(CSOUND *csound)
     arrayvars_localops_init, zak_localops_init,
     scugens_localops_init, emugens_localops_init,
     pvoc_localops_init, spectra_localops_init,
-    vbap_localops_init, structops_localops_init,
+    vbap_localops_init,
     NULL };
 
   const INITFN2 staticmodules2[] = {
